Use of Python
A large dataset of satellite images is taken to study the change in natural resources like forest and water reserves. There is the use of cv2, which is the latest version of OpenCV ( image and video processing library). Apart from it, there are other tools like KMeans and PCA which are also used in this project.